gestao_tatuagens/requirements.txt
asgiref==3.11.1
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==47.0.0
Django==6.0.4
django-allauth==65.16.1
django-cors-headers==4.9.0
django-rest-passwordreset==1.5.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
idna==3.13
pillow==12.2.0
pycparser==3.0
PyJWT==2.12.1
requests==2.33.1
sqlparse==0.5.5
tzdata==2026.2
urllib3==2.6.3
